Refactored code using moment
